Cmd flags to strings, split cleanup from restructure
