+++ dataproc/cleanup/clean_game_description.go
@@ -1,0 +1,65 @@
+package cleanup
+
+import (
+	"github.com/Kaszanas/SC2InfoExtractorGo/datastruct/replay_data"
+	"github.com/icza/s2prot/rep"
+	log "github.com/sirupsen/logrus"
+)
+
+// cleanGameDescription copies the game description,
+// partly verifies the integrity of the data.
+// Has the capability to remove unescessary fields.
+func CleanGameDescription(replayData *rep.Rep) (replay_data.CleanedGameDescription, bool) {
+
+	// Constructing a clean GameDescription without unescessary fields:
+	gameDescription := replayData.InitData.GameDescription
+	gameOptions := gameDescription.GameOptions.Struct
+
+	gameSpeedString := gameDescription.GameSpeed().String()
+
+	isBlizzardMap := gameDescription.IsBlizzardMap()
+	mapAuthorName := gameDescription.MapAuthorName()
+
+	// mapFilename := gameDescription.MapFileName()
+	// log.WithField("mapFilename", mapFilename).Info("Found mapFilename")
+
+	mapFileSyncChecksum := gameDescription.MapFileSyncChecksum()
+
+	mapSizeX := gameDescription.MapSizeX()
+	mapSizeXChecked, okMapSizeX := checkUint32(mapSizeX)
+	if !okMapSizeX {
+		log.WithField("mapSizeX", mapSizeX).
+			Error("Found that value of mapSizeX exceeds uint32")
+		return replay_data.CleanedGameDescription{}, false
+	}
+
+	mapSizeY := gameDescription.MapSizeY()
+	mapSizeYChecked, okMapSizeY := checkUint32(mapSizeY)
+	if !okMapSizeY {
+		log.WithField("mapSizeY", mapSizeY).
+			Error("Found that value of mapSizeY exceeds uint32")
+		return replay_data.CleanedGameDescription{}, false
+	}
+
+	maxPlayers := gameDescription.MaxPlayers()
+	maxPlayersChecked, okMaxPlayers := checkUint8(maxPlayers)
+	if !okMaxPlayers {
+		log.WithField("maxPlayers", maxPlayers).
+			Error("Found that value of maxPlayers exceeds uint8")
+		return replay_data.CleanedGameDescription{}, false
+	}
+
+	cleanedGameDescription := replay_data.CleanedGameDescription{
+		GameOptions:         gameOptions,
+		GameSpeed:           gameSpeedString,
+		IsBlizzardMap:       isBlizzardMap,
+		MapAuthorName:       mapAuthorName,
+		MapFileSyncChecksum: mapFileSyncChecksum,
+		MapSizeX:            mapSizeXChecked,
+		MapSizeY:            mapSizeYChecked,
+		MaxPlayers:          maxPlayersChecked,
+	}
+	log.Info("Defined cleanedGameDescription struct")
+
+	return cleanedGameDescription, true
+}
