+++ dataproc/restructure.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/Kaszanas/SC2InfoExtractorGo/dataproc/cleanup"
 	"github.com/Kaszanas/SC2InfoExtractorGo/datastruct/replay_data"
-	"github.com/icza/s2prot"
 	"github.com/icza/s2prot/rep"
 	log "github.com/sirupsen/logrus"
 )
@@ -18,19 +18,19 @@ func redifineReplayStructure(
 
 	log.Info("Entered redefineReplayStructure()")
 
-	cleanHeader := cleanHeader(replayData)
-	cleanGameDescription, ok := cleanGameDescription(replayData)
+	cleanHeader := cleanup.CleanHeader(replayData)
+	cleanGameDescription, ok := cleanup.CleanGameDescription(replayData)
 	if !ok {
 		return replay_data.CleanedReplay{}, false
 	}
-	cleanInitData, cleanedUserInitDataList, ok := cleanInitData(
+	cleanInitData, cleanedUserInitDataList, ok := cleanup.CleanInitData(
 		replayData,
 		cleanGameDescription)
 	if !ok {
 		return replay_data.CleanedReplay{}, false
 	}
-	cleanDetails, detailsReplayMapField := cleanDetails(replayData)
-	cleanMetadata, metadataReplayMapField := cleanMetadata(replayData)
+	cleanDetails, detailsReplayMapField := cleanup.CleanDetails(replayData)
+	cleanMetadata, metadataReplayMapField := cleanup.CleanMetadata(replayData)
 
 	mapFields := []replay_data.ReplayMapField{
 		detailsReplayMapField,
@@ -49,15 +49,15 @@ func redifineReplayStructure(
 		return replay_data.CleanedReplay{}, false
 	}
 
-	enhancedToonDescMap, ok := cleanToonDescMap(replayData, cleanedUserInitDataList)
+	enhancedToonDescMap, ok := cleanup.CleanToonDescMap(replayData, cleanedUserInitDataList)
 	if !ok {
 		log.Error("Failed to clean toon desc map!")
 		return replay_data.CleanedReplay{}, false
 	}
 
-	messageEventsStructs := cleanMessageEvents(replayData)
-	gameEventsStructs := cleanGameEvents(replayData)
-	trackerEventsStructs := cleanTrackerEvents(replayData)
+	messageEventsStructs := cleanup.CleanMessageEvents(replayData)
+	gameEventsStructs := cleanup.CleanGameEvents(replayData)
+	trackerEventsStructs := cleanup.CleanTrackerEvents(replayData)
 
 	justMessageEvtsErr := replayData.MessageEvtsErr
 	justTrackerEvtsErr := replayData.TrackerEvtsErr
@@ -180,456 +180,4 @@ func adjustMapName(
 	cleanMetadata.MapName = englishMapName
 
 	return true
-}
-
-// cleanHeader copies the header,
-// has the capability of removing unescessary fields.
-func cleanHeader(replayData *rep.Rep) replay_data.CleanedHeader {
-	// Constructing a clean replay header without unescessary fields:
-	elapsedGameLoops := replayData.Header.Loops()
-	// TODO: These values of duration are not verified: https://github.com/icza/s2prot/issues/48
-	// durationNanoseconds := replayData.Header.Duration().Nanoseconds()
-	// durationSeconds := replayData.Header.Duration().Seconds()
-	// version := replayData.Header.Struct["version"].(s2prot.Struct)
-
-	version := replayData.Header.VersionString()
-
-	cleanHeader := replay_data.CleanedHeader{
-		ElapsedGameLoops: uint64(elapsedGameLoops),
-		// DurationNanoseconds: durationNanoseconds,
-		// DurationSeconds:     durationSeconds,
-		Version: version,
-	}
-	log.Info("Defined cleanHeader struct")
-	return cleanHeader
-}
-
-// cleanGameDescription copies the game description,
-// partly verifies the integrity of the data.
-// Has the capability to remove unescessary fields.
-func cleanGameDescription(replayData *rep.Rep) (replay_data.CleanedGameDescription, bool) {
-
-	// Constructing a clean GameDescription without unescessary fields:
-	gameDescription := replayData.InitData.GameDescription
-	gameOptions := gameDescription.GameOptions.Struct
-
-	gameSpeedString := gameDescription.GameSpeed().String()
-
-	isBlizzardMap := gameDescription.IsBlizzardMap()
-	mapAuthorName := gameDescription.MapAuthorName()
-
-	// mapFilename := gameDescription.MapFileName()
-	// log.WithField("mapFilename", mapFilename).Info("Found mapFilename")
-
-	mapFileSyncChecksum := gameDescription.MapFileSyncChecksum()
-
-	mapSizeX := gameDescription.MapSizeX()
-	mapSizeXChecked, okMapSizeX := checkUint32(mapSizeX)
-	if !okMapSizeX {
-		log.WithField("mapSizeX", mapSizeX).
-			Error("Found that value of mapSizeX exceeds uint32")
-		return replay_data.CleanedGameDescription{}, false
-	}
-
-	mapSizeY := gameDescription.MapSizeY()
-	mapSizeYChecked, okMapSizeY := checkUint32(mapSizeY)
-	if !okMapSizeY {
-		log.WithField("mapSizeY", mapSizeY).
-			Error("Found that value of mapSizeY exceeds uint32")
-		return replay_data.CleanedGameDescription{}, false
-	}
-
-	maxPlayers := gameDescription.MaxPlayers()
-	maxPlayersChecked, okMaxPlayers := checkUint8(maxPlayers)
-	if !okMaxPlayers {
-		log.WithField("maxPlayers", maxPlayers).
-			Error("Found that value of maxPlayers exceeds uint8")
-		return replay_data.CleanedGameDescription{}, false
-	}
-
-	cleanedGameDescription := replay_data.CleanedGameDescription{
-		GameOptions:         gameOptions,
-		GameSpeed:           gameSpeedString,
-		IsBlizzardMap:       isBlizzardMap,
-		MapAuthorName:       mapAuthorName,
-		MapFileSyncChecksum: mapFileSyncChecksum,
-		MapSizeX:            mapSizeXChecked,
-		MapSizeY:            mapSizeYChecked,
-		MaxPlayers:          maxPlayersChecked,
-	}
-	log.Info("Defined cleanedGameDescription struct")
-
-	return cleanedGameDescription, true
-}
-
-// cleanInitData copies the init data,
-// partly verifies the integrity of the data.
-// Has the capability to remove unescessary fields.
-func cleanInitData(
-	replayData *rep.Rep,
-	cleanedGameDescription replay_data.CleanedGameDescription) (
-	replay_data.CleanedInitData,
-	[]replay_data.CleanedUserInitData,
-	bool) {
-	// Constructing a clean UserInitData without unescessary fields:
-	var cleanedUserInitDataList []replay_data.CleanedUserInitData
-	for _, userInitData := range replayData.InitData.UserInitDatas {
-		// If the name is an empty string ommit the struct and enter next iteration:
-		name := userInitData.Name()
-		if !(len(name) > 0) {
-			continue
-		}
-
-		combinedRaceLevels := userInitData.CombinedRaceLevels()
-		highestLeague := userInitData.HighestLeague().String()
-		clanTag := userInitData.ClanTag()
-		isInClan := checkClan(clanTag)
-
-		userInitDataStruct := replay_data.CleanedUserInitData{
-			CombinedRaceLevels: uint64(combinedRaceLevels),
-			HighestLeague:      highestLeague,
-			Name:               name,
-			IsInClan:           isInClan,
-			ClanTag:            clanTag,
-		}
-
-		cleanedUserInitDataList = append(
-			cleanedUserInitDataList,
-			userInitDataStruct,
-		)
-	}
-
-	cleanInitData := replay_data.CleanedInitData{
-		GameDescription: cleanedGameDescription,
-	}
-	log.Info("Defined cleanInitData struct")
-	return cleanInitData, cleanedUserInitDataList, true
-}
-
-// cleanDetails copies the details,
-// has the capability of removing unescessary fields.
-func cleanDetails(replayData *rep.Rep) (replay_data.CleanedDetails, replay_data.ReplayMapField) {
-	// Constructing a clean CleanedDetails without unescessary fields
-	detailsGameSpeed := replayData.Details.GameSpeed().String()
-	detailsIsBlizzardMap := replayData.Details.IsBlizzardMap()
-
-	// mapFileName := replayData.Details.MapFileName()
-	// log.WithField("mapFileName", mapFileName).Info("Found mapFileName")
-
-	timeUTC := replayData.Details.TimeUTC()
-	mapNameString := replayData.Details.Title()
-	replayMapField := replay_data.ReplayMapField{
-		MapName: mapNameString,
-	}
-
-	cleanDetails := replay_data.CleanedDetails{
-		GameSpeed:     detailsGameSpeed,
-		IsBlizzardMap: detailsIsBlizzardMap,
-		// PlayerList:    detailsPlayerList, // Information from that part is merged with ToonDescMap
-		// TimeLocalOffset: timeLocalOffset, // This is unused
-		TimeUTC: timeUTC,
-		// MapName: mapNameString, // This is unused
-	}
-	log.Info("Defined cleanDetails struct")
-	return cleanDetails, replayMapField
-}
-
-// cleanMetadata copies the metadata,
-// has the capability of removing unescessary fields.
-func cleanMetadata(
-	replayData *rep.Rep,
-) (replay_data.CleanedMetadata, replay_data.ReplayMapField) {
-	// Constructing a clean CleanedMetadata without unescessary fields:
-	metadataBaseBuild := replayData.Metadata.BaseBuild()
-	metadataDataBuild := replayData.Metadata.DataBuild()
-	// metadataDuration := replayData.Metadata.DurationSec()
-	metadataGameVersion := replayData.Metadata.GameVersion()
-
-	foreignMetadataMapName := replayData.Metadata.Title()
-	mapNameField := replay_data.ReplayMapField{
-		MapName: foreignMetadataMapName,
-	}
-
-	cleanMetadata := replay_data.CleanedMetadata{
-		BaseBuild: metadataBaseBuild,
-		DataBuild: metadataDataBuild,
-		// Duration:    metadataDuration,
-		GameVersion: metadataGameVersion,
-		// Players:     metadataCleanedPlayersList, // This is unused.
-		MapName: foreignMetadataMapName,
-	}
-	log.Info("Defined cleanMetadata struct")
-	return cleanMetadata, mapNameField
-}
-
-// cleanToonDescMap copies the toon description map,
-// changes the structure into a more readable form.
-func cleanToonDescMap(
-	replayData *rep.Rep,
-	cleanedUserInitDataList []replay_data.CleanedUserInitData,
-) (map[string]replay_data.EnhancedToonDescMap, bool) {
-
-	dirtyToonPlayerDescMap := replayData.TrackerEvts.ToonPlayerDescMap
-
-	// Merging data-structures to data.EnhancedToonDescMap
-	enhancedToonDescMap := make(map[string]replay_data.EnhancedToonDescMap)
-	for toonKey, playerDescription := range dirtyToonPlayerDescMap {
-		// Solved https://github.com/Kaszanas/SC2InfoExtractorGo/issues/51
-		// Initializing enhanced map from the dirtyToonPlayerDescMap:
-		initializedToonDescMap := replay_data.EnhancedToonDescMap{
-			PlayerID:            playerDescription.PlayerID,
-			UserID:              playerDescription.UserID,
-			SQ:                  playerDescription.SQ,
-			SupplyCappedPercent: playerDescription.SupplyCappedPercent,
-			StartDir:            playerDescription.StartDir,
-			StartLocX:           playerDescription.StartLocX,
-			StartLocY:           playerDescription.StartLocY,
-		}
-		enhancedToonDescMap[toonKey] = initializedToonDescMap
-
-		// Merging information held in metadata.Players into data.EnhancedToonDescMap
-		enhancedToonDescMap[toonKey] = mergeToonDescMapWithMetadata(
-			replayData,
-			playerDescription,
-			enhancedToonDescMap[toonKey],
-		)
-
-		// Merging information contained in the details part of the replay:
-		var err error
-		enhancedToonDescMap[toonKey], err = mergeToonDescMapWithDetails(
-			replayData,
-			toonKey,
-			enhancedToonDescMap[toonKey],
-		)
-		if err != nil {
-			log.WithField("error", err.Error()).
-				Error("Failed to merge toon desc map with details")
-			return enhancedToonDescMap, false
-		}
-
-		// Merging information contained in the cleanedUserInitDataList:
-		enhancedToonDescMap[toonKey], err = mergeToonDescMapWithInitPlayerList(
-			cleanedUserInitDataList,
-			enhancedToonDescMap[toonKey],
-		)
-		if err != nil {
-			log.WithField("error", err.Error()).
-				Error("Failed to merge toon desc map with init player list")
-			return enhancedToonDescMap, false
-		}
-
-	}
-
-	return enhancedToonDescMap, true
-}
-
-// mergeToonDescMapWithMetadata merges the rep.PlayerDesc with the game Metadata
-// into the EnhancedToonDescMap
-func mergeToonDescMapWithMetadata(
-	replayData *rep.Rep,
-	playerDescription *rep.PlayerDesc,
-	enhancedToonDescMap replay_data.EnhancedToonDescMap,
-) replay_data.EnhancedToonDescMap {
-
-	metadataPlayers := replayData.Metadata.Players()
-	if len(metadataPlayers) == 0 {
-		log.Warn("No players found in metadata!")
-		return enhancedToonDescMap
-	}
-	metadataEnhancedToonDescMap := enhancedToonDescMap
-
-	for _, metadataPlayer := range metadataPlayers {
-
-		metadataPlayerID := metadataPlayer.PlayerID()
-		playerDescriptionPlayerID := playerDescription.PlayerID
-
-		if metadataPlayerID != playerDescriptionPlayerID {
-			continue
-		}
-
-		// Filling out struct fields:
-		// What should be done in case if some of these fields are empty?
-		metadataEnhancedToonDescMap.AssignedRace = metadataPlayer.AssignedRace()
-		metadataEnhancedToonDescMap.SelectedRace = metadataPlayer.SelectedRace()
-		metadataEnhancedToonDescMap.APM = metadataPlayer.APM()
-		metadataEnhancedToonDescMap.MMR = metadataPlayer.MMR()
-		metadataEnhancedToonDescMap.Result = metadataPlayer.Result()
-	}
-	return metadataEnhancedToonDescMap
-}
-
-// mergeToonDescMapWithDetails merges the rep.PlayerDesc with the game Details,
-// into the EnhancedToonDescMap, and returns the EnhancedToonDescMap.
-func mergeToonDescMapWithDetails(
-	replayData *rep.Rep,
-	toonKey string,
-	enhancedToonDescMap replay_data.EnhancedToonDescMap,
-) (replay_data.EnhancedToonDescMap, error) {
-
-	detailsPlayers := replayData.Details.Players()
-
-	if len(detailsPlayers) == 0 {
-		log.Warn("No players found in details!")
-		return enhancedToonDescMap, nil
-	}
-
-	detailsEnhancedToonDescMap := enhancedToonDescMap
-	for _, player := range detailsPlayers {
-		toonString := player.Toon.String()
-
-		// Toon string doesn't match, keep looking for the right player:
-		if toonString != toonKey {
-			continue
-		}
-
-		// Found the right player, fill out the fields:
-		detailsEnhancedToonDescMap.Name = player.Name
-
-		// Checking if previously ran loop populated the Race information
-		// REVIEW: Should this be a full race name?
-		if detailsEnhancedToonDescMap.AssignedRace == "" {
-			raceLetter := player.Race().Letter
-			if raceLetter == 'T' {
-				detailsEnhancedToonDescMap.AssignedRace = "Terr"
-			}
-			if raceLetter == 'P' {
-				detailsEnhancedToonDescMap.AssignedRace = "Prot"
-			}
-			if raceLetter == 'Z' {
-				detailsEnhancedToonDescMap.AssignedRace = "Zerg"
-			}
-		}
-
-		resultMapPlayerResult := map[string]string{
-			"Unknown": "Undecided",
-			"Victory": "Win",
-			"Defeat":  "Loss",
-			"Tie":     "Tie",
-		}
-
-		playerResult := player.Result().String()
-		playerResult = resultMapPlayerResult[playerResult]
-
-		if detailsEnhancedToonDescMap.Result != "" {
-			if detailsEnhancedToonDescMap.Result != playerResult {
-				log.Warn("Player results are different!")
-				return detailsEnhancedToonDescMap, fmt.Errorf("player results are different")
-			}
-		}
-
-		// Result was empty, fill it out:
-		if detailsEnhancedToonDescMap.Result == "" {
-			detailsEnhancedToonDescMap.Result = playerResult
-		}
-
-		// Filling out struct fields:
-		detailsEnhancedToonDescMap.Region = player.Toon.Region().Name
-		detailsEnhancedToonDescMap.Realm = player.Toon.Realm().Name
-		detailsEnhancedToonDescMap.Color.A = player.Color[0]
-		detailsEnhancedToonDescMap.Color.B = player.Color[1]
-		detailsEnhancedToonDescMap.Color.G = player.Color[2]
-		detailsEnhancedToonDescMap.Color.R = player.Color[3]
-		detailsEnhancedToonDescMap.Handicap = player.Handicap()
-		// There should be only one player that fits the unique toon key,
-		// if the player was found and the values were filled out we can exit the loop:
-		break
-	}
-
-	return detailsEnhancedToonDescMap, nil
-}
-
-func mergeToonDescMapWithInitPlayerList(
-	cleanedUserInitDataList []replay_data.CleanedUserInitData,
-	enhancedToonDescMap replay_data.EnhancedToonDescMap,
-) (replay_data.EnhancedToonDescMap, error) {
-
-	if len(cleanedUserInitDataList) == 0 {
-		log.Warn("No players found in cleanedUserInitDataList!")
-		return enhancedToonDescMap, nil
-	}
-
-	// Merging cleanedUserInitDataList information into data.EnhancedToonDescMap:
-	initEnhancedToonDescMap := enhancedToonDescMap
-	for _, initPlayer := range cleanedUserInitDataList {
-
-		toonMapName := initEnhancedToonDescMap.Name
-		initPlayerName := initPlayer.Name
-
-		// The names don't align, keep looking:
-		if !strings.HasSuffix(toonMapName, initPlayerName) {
-			continue
-		}
-
-		// Both names are empty, this cannot be correct:
-		if initEnhancedToonDescMap.Name == "" && initPlayer.Name == "" {
-			log.Error("Both player names are empty in mergeToonDescMapWithInitPlayerList!")
-			return initEnhancedToonDescMap, fmt.Errorf("both player names are empty")
-		}
-
-		// If the name is empty in the EnhancedToonDescMap, fill it out from the initPlayer:
-		if initEnhancedToonDescMap.Name == "" && initPlayer.Name != "" {
-			initEnhancedToonDescMap.Name = initPlayer.Name
-		}
-		initEnhancedToonDescMap.HighestLeague = initPlayer.HighestLeague
-		initEnhancedToonDescMap.IsInClan = initPlayer.IsInClan
-		initEnhancedToonDescMap.ClanTag = initPlayer.ClanTag
-
-		// No need to look any further:
-		break
-	}
-
-	return initEnhancedToonDescMap, nil
-
-}
-
-// cleanMessageEvents copies the message events,
-// has the capability of removing unescessary fields.
-func cleanMessageEvents(replayData *rep.Rep) []s2prot.Struct {
-	// Constructing a clean MessageEvents without unescessary fields:
-	var messageEventsStructs []s2prot.Struct
-	for _, messageEvent := range replayData.MessageEvts {
-		messageEventsStructs = append(messageEventsStructs, messageEvent.Struct)
-	}
-	log.Info("Defined cleanMessageEvents struct")
-	return messageEventsStructs
-}
-
-// cleanGameEvents copies the game events,
-// has the capability of removing unescessary fields.
-func cleanGameEvents(replayData *rep.Rep) []s2prot.Struct {
-	// Constructing a clean GameEvents without unescessary fields:
-	var gameEventsStructs []s2prot.Struct
-	for _, gameEvent := range replayData.GameEvts {
-		gameEventsStructs = append(gameEventsStructs, gameEvent.Struct)
-	}
-	log.Info("Defined cleanGameEvents struct")
-	return gameEventsStructs
-}
-
-// cleanTrackerEvents copies the tracker events,
-// has the capability of removing unescessary fields.
-func cleanTrackerEvents(replayData *rep.Rep) []s2prot.Struct {
-	// Constructing a clean TrackerEvents without unescessary fields:
-	var trackerEventsStructs []s2prot.Struct
-	for _, trackerEvent := range replayData.TrackerEvts.Evts {
-
-		// https://github.com/Kaszanas/SC2InfoExtractorGo/issues/41
-		if trackerEvent.Struct["evtTypeName"] == "PlayerStats" {
-
-			// Get stats:
-			stats := trackerEvent.Struct["stats"].(s2prot.Struct)
-
-			// Get values:
-			foodUsed := stats["scoreValueFoodUsed"].(int64) / 4096
-			foodMade := stats["scoreValueFoodMade"].(int64) / 4096
-
-			// Overwrite values:
-			trackerEvent.Struct["stats"].(s2prot.Struct)["scoreValueFoodUsed"] = foodUsed
-			trackerEvent.Struct["stats"].(s2prot.Struct)["scoreValueFoodMade"] = foodMade
-		}
-
-		trackerEventsStructs = append(trackerEventsStructs, trackerEvent.Struct)
-	}
-	log.Info("Defined cleanTrackerEvents struct")
-	return trackerEventsStructs
 }
